requirements.txt
fastapi==0.95.0
matplotlib==3.7.1
numpy==1.23.1
paho_mqtt==1.6.1
pandas==1.4.3
pydantic==1.9.1
pymongo==4.2.0
scipy==1.9.0
uvicorn==0.18.2
